+++ src/languages/rust_macros.ts
@@ -44,5 +44,4 @@ macro_rules! dbg_mdl_pretty {
         ($(dbg_pretty!($val)),+,)
     };
 }
-
-`
+`;